ui updated for subtitles autosplit

+++ trunk/avidemux/avidemux/ADM_video/ADM_vidSRT.h
@@ -34,10 +34,10 @@
 		int32_t    _U_percent;
 		int32_t    _V_percent;
 
-		char		 *_fontname;
-		char		 *_subname;
-		char 	*_charset;
-  uint8_t _selfAdjustable;
+		char		*_fontname;
+		char		*_subname;
+		char 		*_charset;
+  		uint32_t	_selfAdjustable;
 
 
 	}SUBCONF;
